chore: detailInfoView 파일 분리 및 출석viewcontroller생성
