+++ Features/MOITDetail/MOITDetail/Implement/MOITDetailAttendance/MOITDetailAttandanceViewController.swift
@@ -1,0 +1,12 @@
+//
+//  MOITDetailAttendanceViewController.swift
+//  MOITDetailImpl
+//
+//  Created by 송서영 on 2023/06/10.
+//  Copyright © 2023 chansoo.MOIT. All rights reserved.
+//
+
+import UIKit
+
+final class MOITDetailAttendanceViewController: UIViewController {
+}
